Move and split hooks
